refactor: move mouse over state
