+++ src/states/mouse-over-state.ts
@@ -1,0 +1,26 @@
+import { ref } from '@vue/composition-api';
+
+/**
+ * state & functions for tracking elements whether they are mouse-over or out.
+ * @param disabled
+ * @return {{onMouseOut: onMouseOut, isMouseOver: Ref<HasDefined<S> extends true ? S : RefValue<T>>, onMouseOver: onMouseOver}}
+ */
+export const mouseOverState = (disabled?: boolean) => {
+    const disable = disabled || false;
+    const isMouseOver = ref(false);
+    const onMouseOver = () => {
+        if (!disable && !isMouseOver.value) {
+            isMouseOver.value = true;
+        }
+    };
+    const onMouseOut = () => {
+        if (!disable && isMouseOver.value) {
+            isMouseOver.value = false;
+        }
+    };
+    return {
+        isMouseOver,
+        onMouseOver,
+        onMouseOut,
+    };
+};
