+++ src/util/composition-helpers.ts
@@ -1,6 +1,5 @@
 import {
-    ComponentRenderProxy,
-    computed, getCurrentInstance, onMounted, onUnmounted, ref, Ref,
+    computed, ref, Ref,
 } from '@vue/composition-api';
 
 /**
@@ -75,30 +74,3 @@ export function makeOptionalProxy <T=any>(name: string, vm, initData: any, event
         },
     });
 }
-
-
-/**
- * state & functions for tracking elements whether they are mouse-over or out.
- * @param disabled
- * @return {{onMouseOut: onMouseOut, isMouseOver: Ref<HasDefined<S> extends true ? S : RefValue<T>>, onMouseOver: onMouseOver}}
- */
-export const mouseOverState = (disabled?: boolean) => {
-    const disable = disabled || false;
-    const isMouseOver = ref(false);
-    const onMouseOver = () => {
-        if (!disable && !isMouseOver.value) {
-            isMouseOver.value = true;
-        }
-    };
-    const onMouseOut = () => {
-        if (!disable && isMouseOver.value) {
-            isMouseOver.value = false;
-        }
-    };
-    return {
-        isMouseOver,
-        onMouseOver,
-        onMouseOut,
-    };
-};
-
